+++ stepmania/src/ProfileHtml.cpp
@@ -1003,10 +1003,10 @@ void PrintCaloriesBurned( RageFile &f, const Profile *pProfile, CString sTitle, 
 					Profile::Day day = { when.tm_yday, when.tm_year+1900 };
 					float fDayCals = pProfile->GetCaloriesBurnedForDay( day );
 					fWeekCals += fDayCals;
-					f.Write("<td class='daycalories'><font class='daycalories'>"+((fDayCals==0)?"":Commify(fDayCals))+"</font></td>");
+					f.Write("<td class='daycalories'><font class='daycalories'>"+((fDayCals==0)?"":Commify((int)fDayCals))+"</font></td>");
 					when = AddDays( when, +1 );
 				}
-				f.Write("<td class='weekcalories'><font class='weekcalories'>"+Commify(fWeekCals)+"</font></td>");
+				f.Write("<td class='weekcalories'><font class='weekcalories'>"+Commify((int)fWeekCals)+"</font></td>");
 				f.Write("</tr>");
 				when = AddDays( when, -DAYS_IN_WEEK*2 );
 			}
